feat: fixed admin expire position
