+++ packages/admin/contracts/AdminExpirePosition.sol
@@ -21,9 +21,12 @@
 pragma solidity ^0.8.9;
 
 import { OnlyDolomiteMargin } from "@dolomite-exchange/modules-base/contracts/helpers/OnlyDolomiteMargin.sol";
+import { AccountActionLib } from "@dolomite-exchange/modules-base/contracts/lib/AccountActionLib.sol";
+import { IDolomiteStructs } from "@dolomite-exchange/modules-base/contracts/protocol/interfaces/IDolomiteStructs.sol";
 import { IExpirePositionProxy } from "@dolomite-exchange/modules-base/contracts/proxies/interfaces/IExpirePositionProxy.sol"; // solhint-disable-line max-line-length
 import { AdminRegistryHelper } from "./AdminRegistryHelper.sol";
 import { IDolomiteOwner } from "./interfaces/IDolomiteOwner.sol";
+import { IAdminExpirePosition } from "./interfaces/IAdminExpirePosition.sol";
 
 
 /**
@@ -32,7 +35,7 @@ import { IDolomiteOwner } from "./interfaces/IDolomiteOwner.sol";
  *
  * @notice  Contract that enables an admin to set an expiration for a user's position
  */
-contract AdminExpirePosition is OnlyDolomiteMargin, AdminRegistryHelper {
+contract AdminExpirePosition is OnlyDolomiteMargin, AdminRegistryHelper, IAdminExpirePosition {
 
     // ===================================================================
     // ============================ Constants ============================
@@ -41,18 +44,18 @@ contract AdminExpirePosition is OnlyDolomiteMargin, AdminRegistryHelper {
     bytes32 private constant _FILE = "AdminExpirePosition";
 
     bytes32 public constant ADMIN_EXPIRE_POSITION_ROLE = keccak256("AdminExpirePosition");
-    address public immutable EXPIRE_POSITION_PROXY;
+    address public immutable EXPIRY;
 
     // ===================================================================
     // =========================== Constructors ==========================
     // ===================================================================
 
     constructor(
-        address _expirePositionProxy,
+        address _expiry,
         address _adminRegistry,
         address _dolomiteMargin
     ) OnlyDolomiteMargin(_dolomiteMargin) AdminRegistryHelper(_adminRegistry) {
-        EXPIRE_POSITION_PROXY = _expirePositionProxy;
+        EXPIRY = _expiry;
     }
 
     // ===================================================================
@@ -60,15 +63,34 @@ contract AdminExpirePosition is OnlyDolomiteMargin, AdminRegistryHelper {
     // ===================================================================
 
     function expirePositions(
-        IExpirePositionProxy.ExpirePositionParams[] memory positions
+        ExpirePositionParams[] memory positions
     ) external checkPermission(this.expirePositions.selector, msg.sender) {
+        uint256 positionLen = positions.length;
+        uint256 actionsLen;
 
-        IDolomiteOwner(DOLOMITE_MARGIN_OWNER()).submitTransactionAndExecute(
-            EXPIRE_POSITION_PROXY,
-            abi.encodeWithSelector(
-                IExpirePositionProxy.expirePositions.selector,
-                positions
-            )
-        );
+        IDolomiteStructs.AccountInfo[] memory accounts = new IDolomiteStructs.AccountInfo[](positionLen);
+        for (uint256 i; i < positionLen; ++i) {
+            accounts[i] = positions[i].account;
+            actionsLen += positions[i].owedMarkets.length;
+        }
+
+        IDolomiteStructs.ActionArgs[] memory actions = new IDolomiteStructs.ActionArgs[](actionsLen);
+        uint256 actionsCursor;
+
+        for (uint256 i; i < positionLen; ++i) {
+            ExpirePositionParams memory position = positions[i];
+
+            for (uint256 j; j < position.owedMarkets.length; ++j) {
+                actions[actionsCursor++] = AccountActionLib.encodeExpirationAction(
+                    position.account,
+                    /* accountId = */ i,
+                    position.owedMarkets[j],
+                    address(EXPIRY),
+                    /* _expiryTimeDelta = */ position.expirationTimestamp - block.timestamp
+                );
+            }
+        }
+
+        DOLOMITE_MARGIN().operate(accounts, actions);
     }
 }
